+++ CHANGELOG.md
@@ -1,10 +1,49 @@
 # Formuliser Changelog
 
+#### Changelog Notes
+> Regex used to replace compounds in code to compounds in markdown: `'(.*)': '(.*)'` to find and then using `  - $2 ($1)` to replace
+
+---
+
+## v3.0.0
+
+### Features
+- Added support for reverse elements
+- Added support reverse compounds
+- Added procedural compound name generation (only with 2 elements for now)
+- Updated format of how compounds are shown
+- Added **20** new compounds (6 > 26 total):
+  - Octane (C8H18),
+  - Baking Soda (NaHCO3),
+  - Methane (CH4),
+  - Ammonia (NH3),
+  - Benzene (C6H6),
+  - Ethanol (C2H6O),
+  - Hydrogen Peroxide (H2O2),
+  - Acetic Acid (CH3COOH),
+  - Methanol (CH3OH),
+  - Buckminsterfullerene (C60),
+  - Ethylene (C2H4),
+  - Acetone (C3H6O),
+  - Butane (C4H10),
+  - Citric Acid (C6H8O7),
+  - Tetrahydrofuran (C4H8O),
+  - Hexane (C6H14),
+  - Propane (C3H8),
+  - Vinegar (C2H4O2),
+  - Camphor (C10H16O),
+  - Salt (NaCl)
+
+### Fixes
+- Made information in the bottom right behind everything else
+
+
 ## v2.1.0
 
 ### Features
 - Added version and information to bottom right
 - Added errors
+
 
 ## v2.0.0
 
@@ -22,6 +61,7 @@
     - Flerovium
     - Tennessine
     - Nihonium
+
 
 ## v1.0.0